Rename Anopheles/PerHost files → PerHostAnoph
